Move 'test_fast_readWithElementTree' to the proper file
